NEW: Bert Score
NEW: Utils query now contains a method to load docs by their uuid from ChatNoir. A local cache is used for this.

NEW: Example of the key file needed for ChatNoir and OpenPageRank.

NEW: load_bert.py is responsible for loading a finetuned bert model and extends a dataframe with the score predicted by the model.

NEW: .gitignore to ignore the created cache files and the key.csv.

UPDATED: combiner.py updated to support the bert score.