Addition of the SVM
UPDATE: combiner now has a "dryrun" which trains the svm based on the calculated scores. The preprocessing is bypassed and only the known documents from the qrels are used.

UPDATE: get_title now returns a dataframe with the columns [TopicID,topic].

NEW: svm.py a support vector machine that works based on a Dataframe. In the Dataframe all columns are used as scores, which begin with "Score_".

NEW: from_file loads a saved svm from a given path.

NEW: new creates a new svm.

NEW: save saves the trained svm to allow further use later on.

NEW:train trains the svm based on a Dataframe, only columns starting with "Score_" are considered. The scores are thereby z-transformed.

NEW:predict calculates the prediction of the svm per column and adds a new column "final" in which the results of the svm are stored.

NEW: Function train in svm.py trains the svm based on a Dataframe and stores it in a unique subdirectory.

NEW: Function df_add_score in svm.py. This adds a column final to the given Dataframe, which is calculated based on all known scores from the training. Previously, training must have run for the appropriate combination of scores.