Ties in the final score and new normalization for the SVM
UPDATED: The normalized values are now scaled using a sigmoid function to catch values that deviate strongly from the training data set.

UPDATED: If several documents have the same score at the end, the score of all but one is reduced to the next lower value. This is repeated until there are no more ties.