Added response controls and increased numerical stability.
UPDATED: The default value of the argument test in combiner is now false.

FIX: load_bert no longer terminates the program if an error occurs in bert. Instead, None is entered as prediction and the corresponding text and topic are logged together with the error.

FIX: query now controls the response code. If a response between 400-600 is returned, an error is raised. Subsequently, the already existing exception handling will take effect.

FIX: df2trec no longer uses the smallest possible number but this number times 100 to avoid numerical errors.